requirements.txt
numpy>=1.26.4
pandas>=2.2.3
qdrant-client>=1.7.3
pinecone-text>=0.9.0
python-dotenv>=1.0.1
sentence-transformers>=3.4.1
tqdm>=4.66.5
nltk>=3.8.1
transformers>=4.35.0
Pillow>=10.0.0
diffusers>=0.21.0
torch>=2.0.0
accelerate>=0.21.0
httpx>=0.24.0
asyncio>=3.4.3
